Re-instate missing Prettifier
